+++ core/src/main/kotlin/sift/core/asm/signature/FormalTypeParameterVisitor.kt
@@ -29,4 +29,8 @@ internal class FormalTypeParameterVisitor(
     override fun visitTypeArgument() {
         sv?.visitTypeArgument()
     }
+
+    override fun visitTypeVariable(name: String) {
+        sv?.visitTypeVariable(name)
+    }
 }